Move to new indentation system and create a by-line greyspace/ blackspace handling system.
Indentation tracking is removed node generators' and moved into
tracking the depth of each node; and then computing indentatation
levels based on the node's nesting level.

Indentation is computed based on a simple set of rules, starting
from the current line and going up to the first line:
 1. If current nesting level is less than previous nesting level,
    continue upwards progress, skipping remaining steps.
    This is the pattern of indented lines being skipped on
    subsequent passes.
 2. If current nesting level is equal to the previous nesting level;
    set the indent level to the current indent level and return it.
 3. If current nesting level is greater than the previous nesting
    level; set the indent level to the current indent level + 1,
    and return it. This is the "indent by 1" case.

The nesting level for a given line is decided based on the first
blackspace found on the line.

Certain lines (closing parentheses) are set to have the same
indentation level as their parents to correctly represent the
common case of function bodies: the closing brace should have
equal indentation as the "function" keyword.

Small updates throughout the codebase relate to removing code calling
removed (and obsolete) functions.