Allow generating code that takes an input depending on ASI.
This is potentially dangerous due to it being capable of not throwing when
actually encountering an error. This is especially true in the ensureNewline
method, where if it encounters blackspace instead of greyspace, it assumes
that the generator had just encountered automatic semicolon insertion (ASI),
and force inserts a newline, not advancing the iterator.